requirements.txt
asgiref==3.2.10
Django==3.1
django-crispy-forms==1.9.2
django-money==1.1
django-phone-field==1.8.1
Pillow==7.2.0
py-moneyed==0.8.0
pytz==2020.1
sqlparse==0.3.1
users==1.0.dev0
